refactor: use unit testable real python functions for most templated remote_functions code
